api/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
SQLAlchemy==1.3.24
google-generativeai